community/telegram-bot/requirements.txt

python-telegram-bot==20.7
aiohttp==3.9.1
python-dotenv==1.0.0
asyncio==3.4.3
